project/b-patrushev/requirements.txt
numba==0.55.2
numpy==1.22.2
Pillow==9.1.1
